third_party/WebKit/LayoutTests/fast/frames/sandboxed-iframe-workers.html

<!DOCTYPE html>
<html>
<head>
</head>
<body>
    <script>
        if (window.testRunner) {
            testRunner.dumpAsText();
            testRunner.dumpChildFrames();
        }
    </script>
    <p>This test verifies that shared workers are not available in a sandboxed iframe.</p>
    <iframe sandbox="allow-scripts" srcdoc="
        <script>
            try {
                var s = new SharedWorker('http://127.0.0.1:8080/fast/frames/resources/sandboxed-iframe-workers.js');
                console.log('FAIL: SharedWorker creation should have thrown an exception.');
            } catch (e) {
                console.log('PASS: SharedWorker creation threw \'' + e + '\'.');
            }
        </script>
    "></iframe>
</body>
</head>
</html>
