third_party/blink/web_tests/fast/frames/sandboxed-iframe-workers-expected.txt

CONSOLE MESSAGE: PASS: SharedWorker creation threw 'SecurityError: Failed to construct 'SharedWorker': Access to shared workers is denied to origin 'null'.'.
This test verifies that shared workers are not available in a sandboxed iframe.

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
